Fix Objective-C compiler warnings for protocol conformance and type safety
- Add NSTextViewDelegate protocol conformance to ChatInput class
- Add NSTableViewDataSource and NSTableViewDelegate protocol conformance to ChatConversation class
- Fix integer precision warnings by changing int parameters to NSInteger
- Resolve variable shadowing in MainWindowController delegate methods
- Add conversationHistory getter method to Networking class for proper encapsulation

+++ Sources/ChatConversation/ChatConversation.m
@@ -110,11 +110,11 @@
 
 #pragma mark - NSTableView DataSource
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView {
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView {
     return [messages count];
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(int)row {
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row {
     if (row < [messages count]) {
         ChatMessage *message = [messages objectAtIndex:row];
         return [message description];
